Changed files organisation
