+++ crates/amethyst-vulkan/src/prelude.rs
@@ -19,11 +19,14 @@ pub use crate::{
         LogicalDevice, LogicalDeviceCreateInfo, PhysicalDevice, PhysicalDevicePickInfo,
         RenderDevice, RenderDevicePickInfo,
     },
+    format::Format,
     image::{
         sampler::{ImageSampler, ImageSamplerCreatInfo},
+        surface::{ColorSpace, Surface},
         view::{ImageView, ImageViewCreateInfo, ImageViewKind},
-        Image, ImageAccess, ImageAspectFlags, ImageCreateInfo, ImageDescriptorInfo, ImageFormat,
-        ImageLayout, ImageMemory, ImageSubResourceLayer, ImageSubResourceRange, ImageUsage,
+        Extent2D, Extent3D, Image, ImageAccess, ImageAspectFlags, ImageCreateInfo,
+        ImageDescriptorInfo, ImageFormat, ImageLayout, ImageMemory, ImageSubResourceLayer,
+        ImageSubResourceRange, ImageUsage,
     },
     pipeline::{
         AttachmentLoadOp, AttachmentStoreOp, CullMode, FillMode, FrontFace, Pipeline,
@@ -31,7 +34,6 @@ pub use crate::{
     },
     queue::{Queue, QueueIndex, QueueSubmitInfo},
     shader::{Shader, ShaderCompileInfo, ShaderSource, ShaderSourceType, ShaderStages, ShaderType},
-    surface::{ColorSpace, Extent2D, Extent3D, Format, Surface},
     swapchain::{
         PresentMode, Swapchain, SwapchainCreatInfo, SwapchainFormat, SwapchainPresentInfo,
         SwapchainSupport,
